Trying to clean up a bit
